Move CenterCropOrPad to crop_or_pad module
